chore: break down babel api tests into multiple files
