+++ packages/babel-plugin/src/__tests__/expression-evaluation.test.tsx
@@ -26,6 +26,19 @@ describe('import specifiers', () => {
     expect(actual).toInclude('.cc-hash-test{font-size:16px}');
   });
 
+  it('should inline mutable identifier that is not mutated', () => {
+    const actual = transform(`
+      import '@compiled/core';
+      import React from 'react';
+
+      let notMutatedAgain = 20;
+
+      <div css={{ fontSize: notMutatedAgain }}>hello world</div>
+    `);
+
+    expect(actual).toInclude('.cc-hash-test{font-size:20px}');
+  });
+
   it('should bail out evaluating expression referencing a mutable identifier', () => {
     const actual = transform(`
       import '@compiled/core';
